refactor: Migrate LoadingButton to MaterialLoadingButton
